Create WhiteLineDetectionV5
This code takes only lower half, does HSV filtering and Canny Edge detection and then fuse them by using AND operation. It allows for tweaking the values and selecting appropriate HSV thresholds and Canny edge thresholds. In this code we also have Gaussian blur, this helps you to reduce the detail so that edge detection's noise is removed.